Accumulator: add terraform EMR description.
Adding a EMR terraform deployment in charge running the Hadoop
services.

We're spinning a VPC with two subnetworks: a public one and a private
one. Public IP addresses are allocated to the EC2 machines created on
the public network. Private adresses are allocated to the EC2
instances created on the private subnetwork. Gateway-wise, we NAT the
egress coming from the private network, we just forward the egress
coming out of the public sub-network.

The Sequencer is allocated on the public network, the Accumulator
on the private network. The Sequencer is able to connect to the master
Hadoop node through the VPC.

We're missing the sequencer here, I'm still not 100% sure how to run
it (podman on EC2?) and how to connect it to the EMR cluster (likely
through a IAM policy?).