EMR: revise IAM policies
We used to define custom IAM policies for the EMR cluster. While it
kept the project self-contained (the EMR-specifc generic IAM policies
sadly need a AWS cli call to be bootstrapped), it would have been more
tricky to maitain on the long run to keep the alignment with the EMR
cluster.

We roll back to use the default EMR policies for the cluster. We add
some custom policies of top of the default ones to get access to the
S3 buckets.