Infra: add Sequencer
We create a EC2 machine running the fast-zk-rollup docker container.

The NixOS EC2 machine loads this container through a systemd-managed
podman runtime. The EC2 machine also runs a Nginx reverse proxy
listening on 80. We'll add a TLS layer once we'll have a proper
route53 setup for the domain name.

We attach a policy to the EC2 machine allowing the services running on
this machine to retrieve a temporary access token to control the EMR
setup (ie. adding tasks, etc.).