Infra: open the floodgates to EMR cluster
Temporarily adding the hadoop machines to the public subnet, the one
in which machines are getting assigned public IPs.

The default security group for the hadoop machines is already full
open (they were supposed to be run in a private subnet), meaning we'll
get full internet access to them.

We're also adding ycryptx's ssh keypair to the EMR machines to allow
him to directly ssh to them.