Infra: add Sequencer to private EMR network
Somehow overlooked this during the first cluster deployment. Strange
we did not stumbled upon this annoyance earlier… I assume we only used
the EMR API to interact with the cluster so far, it explains why.

Adding a new network interface to the sequencer. This network
interface is connected to the private EMR network.

Relaxing a bit the dev security group as well. Allowing us, among
other things, to ICMP ping the EMR nodes.

Finally, allowing the sequencer to describe the EC2 instances. This is
definitely not a thing we want to have in production, however, this is
something **really** useful for the dev flow.