Infra: open the floodgates to EMR cluster (#29)
* Infra: open the floodgates to EMR cluster

Temporarily adding the hadoop machines to the public subnet, the one
in which machines are getting assigned public IPs.

The default security group for the hadoop machines is already full
open (they were supposed to be run in a private subnet), meaning we'll
get full internet access to them.

We're also adding ycryptx's ssh keypair to the EMR machines to allow
him to directly ssh to them.

* Infra: create dev sg + keypair

These ressources are interactively used from the AWS CLI.

* Infra: add Sequencer to private EMR network

Somehow overlooked this during the first cluster deployment. Strange
we did not stumbled upon this annoyance earlier… I assume we only used
the EMR API to interact with the cluster so far, it explains why.

Adding a new network interface to the sequencer. This network
interface is connected to the private EMR network.

Relaxing a bit the dev security group as well. Allowing us, among
other things, to ICMP ping the EMR nodes.

Finally, allowing the sequencer to describe the EC2 instances. This is
definitely not a thing we want to have in production, however, this is
something **really** useful for the dev flow.